% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Conn_util_funcs.R
\name{calc_linex_a}
\alias{calc_linex_a}
\title{estimate optimal 'a' parameter for linex loss function}
\usage{
calc_linex_a(Pred.G, Obs.G, min.a = 1e-05, max.a = 1)
}
\arguments{
\item{Pred.G}{Predicted group abundance}

\item{Obs.G}{Observed group abundance}

\item{min.a}{Minimum value for linex 'a' parameter}

\item{max.a}{Maximum value for linex 'a' parameter}
}
\value{
The optimal tuning parameter for linex loss function as determined by minimum sum of squares
}
\description{
estimate optimal 'a' parameter for linex loss function
}
\author{
Paul B. Conn
}
\keyword{linex}

