% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Conn_util_funcs.R
\name{d_logP_omega}
\alias{d_logP_omega}
\title{Calculate partial derivative vector for for Langevin omega MH updates (for CPIF model)
(Assuming function is called separately for each multinomial update)}
\usage{
d_logP_omega(Omega, Counts, Mu, tau)
}
\arguments{
\item{Omega}{Current values of Omega}

\item{Counts}{Animal counts for current time step being updated}

\item{Mu}{Expected value for Omega}

\item{tau}{precision for exchangable errors}
}
\value{
Vector of partial derivates evaluated at Omega
}
\description{
Calculate partial derivative vector for for Langevin omega MH updates (for CPIF model)
(Assuming function is called separately for each multinomial update)
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{expit}
\keyword{logit,}

