% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Conn_util_funcs.R
\name{expit}
\alias{expit}
\title{inverse logit transformation}
\usage{
expit(x)
}
\arguments{
\item{x}{quantity to be transformed}
}
\value{
real valued quantities
}
\description{
inverse logit transformation
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{expit}
\keyword{logit,}

