% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clim_proj_format_funcs.R
\name{format_climate}
\alias{format_climate}
\title{Produce a climate dataframe for sageAbundance simulations}
\usage{
format_climate(tdata, pdata, years)
}
\arguments{
\item{tdata}{subset of temperature dataframe for a single model and scenario}

\item{pdata}{subset of precipitation dataframe for a single model and scenario}

\item{years}{vector of desired years in output}
}
\value{
clim_data formatted climate dataframe for simulating population model
}
\description{
Produce a climate dataframe for sageAbundance simulations
}
\author{
Andrew Tredennick
}

