% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Conn_util_funcs.R
\name{linear_adj_RW2}
\alias{linear_adj_RW2}
\title{Produce an RW2 structure matrix for a line (e.g. for time series)}
\usage{
linear_adj_RW2(x)
}
\arguments{
\item{x}{length of vector}
}
\value{
RW2 precision matrix
}
\description{
Produce an RW2 structure matrix for a line (e.g. for time series)
}
\author{
Paul Conn
}
\keyword{adjacency}

