% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stanfit_functions.R
\name{model_nocovars}
\alias{model_nocovars}
\title{Estimate model parameters in GLMM using rstan with no environmental covariates}
\usage{
model_nocovars(y, lag, K, cellid, iters = 2000, warmup = 1000,
  nchains = 1)
}
\arguments{
\item{y}{Vector of responses (cover observations)}

\item{lag}{Vector of lag response for temporal response (previous year's cover)}

\item{K}{Spatial field matrix}

\item{cellid}{Vector of unique ID's for each spatial cell, replicate through time}

\item{iters}{Number of MCMC iterations to run}

\item{warmup}{Number of MCMC iterations to throw out before sampling (< iters)}

\item{nchains}{Number of MCMC chains to sample}
}
\value{
ggs object with MCMC values from rstan
}
\description{
Estimate model parameters in GLMM using rstan with no environmental covariates
}
\author{
Andrew Tredennick
}

