% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Conn_util_funcs.R
\name{multinom_logL_inits}
\alias{multinom_logL_inits}
\title{neg log likelihood function for generating initial estimates of Beta to start off Markov chain}
\usage{
multinom_logL_inits(Beta, Counts, P, X)
}
\arguments{
\item{Beta}{parameter values}

\item{Counts}{Animal counts in sampled cells}

\item{P}{probability of seeing an animal by cell}

\item{X}{model matrix}
}
\value{
multinomial log likelihood
}
\description{
neg log likelihood function for generating initial estimates of Beta to start off Markov chain
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{intial}
\keyword{likelihood}
\keyword{log}
\keyword{multinomial,}
\keyword{values,}

