% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Conn_util_funcs.R
\name{stack_list}
\alias{stack_list}
\title{stack list into matrix}
\usage{
stack_list(L)
}
\arguments{
\item{L}{list vector of matrices}
}
\value{
X one matrix consisting of stacked elements of L
}
\description{
stack list into matrix
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}

