% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Conn_util_funcs.R
\name{stack_list_vector}
\alias{stack_list_vector}
\title{stack list of vectors into one big vector}
\usage{
stack_list_vector(L)
}
\arguments{
\item{L}{list vector of matrices}
}
\value{
X one vector consisting of stacked elements of L
}
\description{
stack list of vectors into one big vector
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}

